requirements_streamlit.txt
streamlit
plotly
boto3
python-dotenv
pandas==1.3.5
numpy==1.21.2
scikit-learn==0.24.2
praw
textblob
datetime